% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf_lag_select.R
\name{adf_lag_select}
\alias{adf_lag_select}
\title{Identify Optimal Lag Order Selection for (Augmented) Dickey-Fuller Tests}
\usage{
adf_lag_select(x, min_lag = 0, max_lag = NULL)
}
\arguments{
\item{x}{a vector}

\item{min_lag}{the minimum lag order to use. Defaults to 0.}

\item{max_lag}{the maximum lag order to use. Defaults to Schwert's (1989) upper lag.}
}
\value{
\code{adf_lag_select()} returns a list of length 3. The first element
in the list is a data frame of a series of (Augmented) Dickey-Fuller tests
for no-drift, no-trend. The second is a data frame of a series of (Augmented)
Dickey-Fuller tests for drift, no trend. The third is a data frame of a series
of (Augmented) Dickey-Fuller tests for a drift and trend. Each data frame has
six columns. The first communicates the particular lag order. The second
communicates the (Augmented) Dickey-Fuller test statistic (tau) for that lag
order. The third and fourth communicate the Akaike and Bayesian information
criteria for the test (respectively). The fifth communicates the absolute
value of the last, lagged first difference (where applicable). The sixth
column contains a note indicating if the lag was suggested by Schwert (1989)
or Said and Dickey (1984).
}
\description{
\code{adf_lag_select()} runs a series of (Augmented) Dickey-Fuller
tests and returns information that may (or may not) be useful in identify
a potential lag order for unit root tests.
}
\details{
This function removes missing values from the vector before calculating test
statistics.

The lower bound lag order suggested by Schwert (1989) and the default suggested
by Said and Dickey (1984) do not meaningfully separate from each other until
the length of the series reaches 127. Under those conditions, if the \code{note}
column returned by this function for a finite series does not identify the
Said and Dickey (1984) default, but does identify the Schwert (1989) lower
bound, interpret the latter as the former.
}
\examples{

adf_lag_select(tbills$tb3m)

}
\author{
Steven V. Miller
}
