% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbills}
\alias{tbills}
\title{Daily maturity rates for U.S. Treasury Bills}
\format{
A data frame with 17,741 observations on the following 4 variables.
\describe{
\item{\code{date}}{a date}
\item{\code{tb3m}}{the three-month treasury bill rate}
\item{\code{tb6m}}{the six-month treasury bill rate}
\item{\code{tb1y}}{the one-year treasury bill rate}
}
}
\usage{
tbills
}
\description{
A data frame on daily (when applicable/available) U.S. Treasury Bill rates.
These are the yield received for investing in a government-issued treasury
security that has a maturity of a given period of time (three months, six
months, or a year).
}
\details{
Data come by way of \pkg{fredr} call. The one-year (52-week) treasury
bill rate was discontinued in 2001 and re-introduced in 2008. Be mindful of
that gap in the series.
}
\keyword{datasets}
