% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USDSEK}
\alias{USDSEK}
\title{The USD/SEK Exchange Rate}
\format{
A data frame with 3905 observations on the following 2 variables.
\describe{
\item{\code{date}}{a date}
\item{\code{close}}{the exchange rate at the close of trading}
}
}
\usage{
USDSEK
}
\description{
A data frame on the USD/SEK exchange rate (i.e. how many Swedish crowns does
one dollar get you).
}
\details{
Data come by way of \pkg{quantmod}.
}
\keyword{datasets}
