% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ur_summary.R
\name{ur_summary}
\alias{ur_summary}
\title{Summarize Unit Root Test Simulations}
\usage{
ur_summary(obj, pp_stat = "tau", ...)
}
\arguments{
\item{obj}{the object to be summarized, of class 'spp_test'}

\item{pp_stat}{the statistic to be summarized: either "tau" or "rho"}

\item{...}{additional argument, currently ignored}
}
\value{
\code{ur_summary()} produces console output that offers a summary
assessment about the presence of a unit root based on your simulations.
}
\description{
\code{ur_summary()} provides a summary of the unit root tests
included in this package.
}
\details{
This function makes ample use of the "attributes" element in the
list produced by the unit root simulations.
}
\author{
Steven V. Miller
}
